move deeplink handling to util
