--- Editor/BoltSDK.Editor.asmdef
@@ -1,18 +1,0 @@
-{
-    "name": "BoltSDK.Editor",
-    "rootNamespace": "BoltSDK.Editor",
-    "references": [
-        "BoltSDK"
-    ],
-    "includePlatforms": [
-        "Editor"
-    ],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
